move npmw install task to java:node
